Rewrite for randomForestPredictNoTrim
I completed the first rewrite for randomForestPredictNoTrim. Also
finally tracked down the bug which was halting execution of original
random forest, fixed that on ram box, and added a fixed version here.
It was looking for a value called ‘unknown’ as a classification, which,
in the file, is listed as a blank or ‘’.